+++ src/test/psdReader.spec.ts
@@ -69,8 +69,8 @@ describe('PsdReader', () => {
 
 	// skipping "pattern" test because it requires zip cimpression of patterns
 	// skipping "cmyk" test because we can't convert CMYK to RGB
-	fs.readdirSync(readFilesPath).filter(f => !/pattern|cmyk|ignore-text-align|ignore-missing-object|ignore-text-curve|smart-object-curves/.test(f)).forEach(f => {
-		// fs.readdirSync(readFilesPath).filter(f => /smart-object-curves/.test(f)).forEach(f => {
+	fs.readdirSync(readFilesPath).filter(f => !/pattern|cmyk|ignore-text-align|ignore-missing-object|ignore-text-curve/.test(f)).forEach(f => {
+		// fs.readdirSync(readFilesPath).filter(f => /blend-if/.test(f)).forEach(f => {
 		it(`reads PSD file (${f})`, () => {
 			const basePath = path.join(readFilesPath, f);
 			const fileName = fs.existsSync(path.join(basePath, 'src.psb')) ? 'src.psb' : 'src.psd';
@@ -211,7 +211,7 @@ describe('PsdReader', () => {
 	});
 
 	fs.readdirSync(readWriteFilesPath).forEach(f => {
-		// fs.readdirSync(readWriteFilesPath).filter(f => /smart-filters$/.test(f)).forEach(f => {
+		// fs.readdirSync(readWriteFilesPath).filter(f => /blend-if$/.test(f)).forEach(f => {
 		it(`reads-writes PSD file (${f})`, () => {
 			const ext = fs.existsSync(path.join(readWriteFilesPath, f, 'src.psb')) ? 'psb' : 'psd';
 			const psd = readPsdFromFile(path.join(readWriteFilesPath, f, `src.${ext}`), {
@@ -237,7 +237,7 @@ describe('PsdReader', () => {
 			// fs.writeFileSync('temp2.txt', require('util').inspect(psd2, false, 99, false), 'utf8');
 
 			const expected = fs.readFileSync(path.join(readWriteFilesPath, f, `expected.${ext}`));
-			compareBuffers(actual, expected, `read-write-${f}`, 0x0);
+			compareBuffers(actual, expected, `read-write-${f}`, 0x6140);
 		});
 	});